create sitemap.js, call sitemap in redirect-list.js, call redirect-list.js in server(start.js)
